refactor : 개인 채팅 리팩토링
