+++ src/pages/directChat/utils.js
@@ -1,0 +1,15 @@
+export const getReceiver = (userId, chatRoomInfo) => {
+  let receiverId;
+  let receiverName;
+  if (chatRoomInfo.createMemberId === userId) {
+    receiverId = chatRoomInfo.joinMemberId;
+    receiverName = chatRoomInfo.joinMemberName;
+  } else {
+    receiverId = chatRoomInfo.createMemberId;
+    receiverName = chatRoomInfo.createMemberName;
+  }
+  return {
+    receiverId,
+    receiverName,
+  };
+};
